remove redundant cast

+++ src/Gpu/Device/Emulator/RuntimeManager/handleRead.cpp
@@ -12,7 +12,7 @@ namespace Gpu::Device {
         const bool ok = memory_.read(msg.ptr, buffer);
         return {
             ok ? Ipc::FailureCodes::IpcSuccess : Ipc::FailureCodes::ReadError,
-            ok ? static_cast<uint64_t>(buffer.size()) : 0,
+            ok ? buffer.size() : 0,
             ok ? std::move(buffer) : Common::Buffer8{}
         };
     }
